refactor: move StoreDir to its own mod
